Move NU and Stake to currency.py
