+++ nucypher/blockchain/eth/actors.py
@@ -36,7 +36,8 @@ from nucypher.blockchain.eth.registry import AllocationRegistry
 from nucypher.blockchain.eth.utils import (datetime_to_period,
                                            validate_stake_amount,
                                            validate_locktime,
-                                           calculate_period_duration, Stake, NU)
+                                           calculate_period_duration)
+from nucypher.blockchain.eth.currency import NU, Stake
 
 
 def only_me(func):
